+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+# v3.4.0
+
+- [added] `firebase.App` now provides a new `DatabaseWithURL()` function
+  for initializing a database client from a URL.
+
 # v3.3.0
 
 - [fixed] Fixing a regression introduced in 3.2.0, where `VerifyIDToken()`